Update README.md
This project is part of the Frontend Mentor challenge, which aims to create a responsive Four Card Feature Section. In this challenge, you will create a layout with four cards that display an image, title, and a short description of the feature or service offered. These cards must be able to adapt to different screen sizes (responsive), and have an interactive effect when hovered by the user.